% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESM262Final-package.R
\docType{package}
\name{ESM262Final-package}
\alias{ESM262Final}
\alias{ESM262Final-package}
\title{ESM262Final: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
